requirements.txt
PySide6
python-vlc
requests
